requirements.txt
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
blinker==1.8.2
boto3==1.34.134
botocore==1.34.134
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.3.0
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
fpdf==1.7.2
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
lxml==5.2.2
MarkupSafe==2.1.5
openai==1.35.7
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.3
zipp==3.23.0
